Refactor 8 components to stay under 1000-line limit
After doc tests were added in PR #428, 8 component files exceeded the
project's 1000-line limit. This refactor splits each file by extracting
the large state impl blocks into separate `state.rs` (or `builders.rs`
for chart) files. No functionality is changed — only file organization.

Files split:
- step_indicator/mod.rs: 1149 -> 622 lines (state.rs: 540)
- conversation_view/mod.rs: 1112 -> 324 lines (state.rs: 802)
- chart/mod.rs: 1077 -> 655 lines (builders.rs: 433, existing state.rs preserved)
- table/mod.rs: 1042 -> 427 lines (state.rs: 625)
- histogram/mod.rs: 1042 -> 463 lines (state.rs: 590)
- box_plot/mod.rs: 1036 -> 635 lines (state.rs: 412)
- data_grid/mod.rs: 1028 -> 560 lines (state.rs: 480)
- usage_display/mod.rs: 1016 -> 598 lines (state.rs: 428)

All 2299 doc tests pass and clippy is warning-free.
